requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
opentelemetry-api>=1.24
opentelemetry-sdk>=1.24
opentelemetry-instrumentation-fastapi>=0.45b0